Machine mode: respond to commands to signal success
After processing a machine-mode instruction, respond with "::ok\r\n".
This way the host side knows if an instruction was processed or not.

Also instructions can carry an optional "sequence number" now, which
will be echoed in the "ok" response.

Example:
  ::lightnode:23[...]
responds with:
  ::ok:23

The sequence number can be used on the host side to synchronize
specific function calls with their response.

Unknown instructions respond with
  ::error
instead.